refactor: :ambulance: Distribucion de carpetas v1
Se esta organizando por carpetas el codigo realizado en dias anteriores
