+++ controllers/manejo_usuarios/controller.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { PORT, HOST } from "./config.js";
+import { PORT, HOST } from "../../config.js";
 
 export const renderLogin = (req, res) => {
     res.render("manejo_usuarios/login");
